stock_market/requirements.txt
finnhub-python
tornado
yoyo-migrations
psycopg2
asyncpg
